Create new snooze settings menu
