+++ app/src/main/res/drawable/snooze.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:height="16dp"
+    android:width="16dp"
+    android:viewportHeight="960"
+    android:viewportWidth="960">
+      
+    <path android:fillColor="?android:colorForeground" android:pathData="M384,624h192v-48L459,576l117,-96v-48L384,432v48h117l-117,96v48ZM348.87,837.4q-61.14,-26.6 -106.4,-71.87 -45.27,-45.26 -71.87,-106.4Q144,598 144,528t26.6,-131.13q26.6,-61.14 71.87,-106.4 45.26,-45.27 106.4,-71.87Q410,192 480,192t131.13,26.6q61.14,26.6 106.4,71.87 45.27,45.26 71.87,106.4Q816,458 816,528t-26.6,131.13q-26.6,61.14 -71.87,106.4 -45.26,45.27 -106.4,71.87Q550,864 480,864t-131.13,-26.6ZM481,521ZM237,115l51,51 -170,170 -51,-51 170,-170ZM723,115 L893,285 842,336 672,166 723,115ZM479.78,792Q590,792 667,715.22t77,-187Q744,418 667.22,341t-187,-77Q370,264 293,340.78t-77,187Q216,638 292.78,715t187,77Z"/>
+    
+</vector>
