+++ app/src/main/res/values-bg/strings.xml
@@ -289,4 +289,8 @@
     <string name="location_snooze_no_play_services">Отлагането по местоположение изисква Google Play Services, което не е налично на това устройство</string>
     <string name="location_snooze_no_home_location">Моля, задайте домашно местоположение преди да активирате отлагането по местоположение</string>
     <string name="location_snooze_background_permission">За повторно известяване при завръщане вкъщи, разрешете на MedTimer достъп до местоположението на заден план</string>
+    <string name="snooze_settings">Настройки за отлагане</string>
+    <string name="snooze_settings_summary">Продължителност на отлагането и отлагане по местоположение</string>
+    <string name="home_location">Домашно местоположение</string>
+    <string name="home_location_not_set">Не е зададено</string>
 </resources>
