+++ app/src/main/res/values-da/strings.xml
@@ -289,4 +289,8 @@
     <string name="location_snooze_no_play_services">Placeringsbaseret snooze kræver Google Play Services, som ikke er tilgængeligt på denne enhed</string>
     <string name="location_snooze_no_home_location">Angiv en hjemplacering, inden du aktiverer placeringsbaseret snooze</string>
     <string name="location_snooze_background_permission">For at blive adviseret ved hjemkomst, tillad MedTimer at tilgå placeringen i baggrunden</string>
+    <string name="snooze_settings">Snooze-indstillinger</string>
+    <string name="snooze_settings_summary">Slumrevarighed og placeringsbaseret snooze</string>
+    <string name="home_location">Hjemsted</string>
+    <string name="home_location_not_set">Ikke angivet</string>
 </resources>
