+++ app/src/main/res/values-de/strings.xml
@@ -289,4 +289,8 @@
     <string name="location_snooze_no_play_services">Standortbasiertes Schlummern erfordert Google Play Services, die auf diesem Gerät nicht verfügbar sind</string>
     <string name="location_snooze_no_home_location">Bitte legen Sie einen Heimatstandort fest, bevor Sie das standortbasierte Schlummern aktivieren</string>
     <string name="location_snooze_background_permission">Um bei der Heimkehr erneut benachrichtigt zu werden, erlauben Sie MedTimer den Hintergrundstandortzugriff</string>
+    <string name="snooze_settings">Snooze-Einstellungen</string>
+    <string name="snooze_settings_summary">Schlummerdauer und standortbasiertes Schlummern</string>
+    <string name="home_location">Heimstandort</string>
+    <string name="home_location_not_set">Nicht festgelegt</string>
 </resources>
