+++ app/src/main/res/values-es/strings.xml
@@ -294,4 +294,8 @@
     <string name="location_snooze_no_play_services">La posposición según ubicación requiere Google Play Services, que no está disponible en este dispositivo</string>
     <string name="location_snooze_no_home_location">Por favor, establezca una ubicación de casa antes de activar la posposición según ubicación</string>
     <string name="location_snooze_background_permission">Para recibir notificaciones al llegar a casa, permita que MedTimer acceda a la ubicación en segundo plano</string>
+    <string name="snooze_settings">Configuración de posposición</string>
+    <string name="snooze_settings_summary">Duración al posponer y posposición según ubicación</string>
+    <string name="home_location">Ubicación de casa</string>
+    <string name="home_location_not_set">No establecida</string>
 </resources>
