+++ app/src/main/res/values-is/strings.xml
@@ -187,4 +187,8 @@
     <string name="location_snooze_no_play_services">Staðsetningarbundið snooze krefst Google Play Services, sem er ekki tiltækt á þessum tæki</string>
     <string name="location_snooze_no_home_location">Vinsamlegast stilltu heimastaðsetningu áður en þú virkjar staðsetningarbundið snooze</string>
     <string name="location_snooze_background_permission">Til að fá tilkynningu við heimkomu skaltu leyfa MedTimer að fá aðgang að staðsetningu í bakgrunni</string>
+    <string name="snooze_settings">Snooze-stillingar</string>
+    <string name="snooze_settings_summary">Blund-tími og staðsetningarbundið snooze</string>
+    <string name="home_location">Heimilisstaðsetning</string>
+    <string name="home_location_not_set">Ekki stillt</string>
 </resources>
