+++ app/src/main/res/values-nl/strings.xml
@@ -289,4 +289,8 @@
     <string name="location_snooze_no_play_services">Locatiegebaseerd sluimeren vereist Google Play Services, die niet beschikbaar zijn op dit apparaat</string>
     <string name="location_snooze_no_home_location">Stel een thuislocatie in voordat u locatiegebaseerd sluimeren inschakelt</string>
     <string name="location_snooze_background_permission">Sta MedTimer toe op de achtergrond toegang te hebben tot de locatie om opnieuw te melden bij thuiskomst</string>
+    <string name="snooze_settings">Sluimerinstellingen</string>
+    <string name="snooze_settings_summary">Sluimertijd en locatiegebaseerd sluimeren</string>
+    <string name="home_location">Thuislocatie</string>
+    <string name="home_location_not_set">Niet ingesteld</string>
 </resources>
