+++ app/src/main/res/values-ru/strings.xml
@@ -296,4 +296,8 @@
     <string name="location_snooze_no_play_services">Отсрочка по местоположению требует Google Play Services, которые недоступны на этом устройстве</string>
     <string name="location_snooze_no_home_location">Пожалуйста, установите домашнее местоположение перед включением отсрочки по местоположению</string>
     <string name="location_snooze_background_permission">Чтобы получать уведомления при возвращении домой, разрешите MedTimer доступ к местоположению в фоне</string>
+    <string name="snooze_settings">Настройки отложенных уведомлений</string>
+    <string name="snooze_settings_summary">Время откладывания и откладывание по местоположению</string>
+    <string name="home_location">Домашнее местоположение</string>
+    <string name="home_location_not_set">Не задано</string>
 </resources>
