+++ app/src/main/res/values-tr/strings.xml
@@ -289,4 +289,8 @@
     <string name="location_snooze_no_play_services">Konuma dayalı erteleme, bu cihazda bulunmayan Google Play Services gerektirir</string>
     <string name="location_snooze_no_home_location">Konuma dayalı ertelemeyi etkinleştirmeden önce lütfen bir ev konumu belirleyin</string>
     <string name="location_snooze_background_permission">Eve gelince yeniden bildirim almak için MedTimer\'ın arka planda konuma erişmesine izin verin</string>
+    <string name="snooze_settings">Erteleme ayarları</string>
+    <string name="snooze_settings_summary">Erteleme süresi ve konuma dayalı erteleme</string>
+    <string name="home_location">Ev konumu</string>
+    <string name="home_location_not_set">Ayarlanmadı</string>
 </resources>
