+++ app/src/main/res/values-uk/strings.xml
@@ -299,4 +299,8 @@
     <string name="location_snooze_no_play_services">Відкладення за місцезнаходженням вимагає Google Play Services, які недоступні на цьому пристрої</string>
     <string name="location_snooze_no_home_location">Будь ласка, встановіть домашнє місцезнаходження перед увімкненням відкладення за місцезнаходженням</string>
     <string name="location_snooze_background_permission">Щоб отримувати сповіщення при поверненні додому, дозвольте MedTimer доступ до місцезнаходження у фоновому режимі</string>
+    <string name="snooze_settings">Налаштування відкладення</string>
+    <string name="snooze_settings_summary">Тривалість відкладення та відкладення за місцезнаходженням</string>
+    <string name="home_location">Домашнє місцезнаходження</string>
+    <string name="home_location_not_set">Не задано</string>
 </resources>
