+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -284,4 +284,8 @@
     <string name="location_snooze_no_play_services">基於位置的小睡需要 Google Play 服務，該服務在此裝置上不可用</string>
     <string name="location_snooze_no_home_location">請在啟用基於位置的小睡之前設定家庭位置</string>
     <string name="location_snooze_background_permission">要在回家時重新接收提醒，請允許 MedTimer 在背景存取位置</string>
+    <string name="snooze_settings">小睡設定</string>
+    <string name="snooze_settings_summary">小睡時間和基於位置的小睡</string>
+    <string name="home_location">家庭位置</string>
+    <string name="home_location_not_set">未設定</string>
 </resources>
