+++ app/src/main/res/values/strings.xml
@@ -204,6 +204,10 @@
     <string name="hide_med_name">Hide medication name in notification</string>
     <string name="re_raise_event">Re-raise event</string>
     <string name="notification_reminder_settings">Notification and reminder settings</string>
+    <string name="snooze_settings">Snooze settings</string>
+    <string name="snooze_settings_summary">Snooze duration and location snooze</string>
+    <string name="home_location">Home location</string>
+    <string name="home_location_not_set">Not set</string>
     <string name="display_settings">Display settings</string>
     <string name="use_relative_date_time_summary">Show relative date/time for reminders</string>
     <string name="use_relative_date_time">Relative date/time</string>
