+++ app/src/main/res/xml/snooze_settings.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <ListPreference
+        android:defaultValue="15"
+        android:entries="@array/snooze_duration"
+        android:entryValues="@array/snooze_duration_values"
+        android:key="snooze_duration"
+        android:title="@string/snooze_duration"
+        app:useSimpleSummaryProvider="true" />
+    <SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:summary="@string/location_snooze_summary"
+        android:title="@string/location_snooze"
+        app:singleLineTitle="false"
+        app:key="location_snooze_enabled" />
+    <Preference
+        android:title="@string/home_location"
+        android:summary="@string/home_location_not_set"
+        app:key="home_location" />
+</androidx.preference.PreferenceScreen>
